#16 Improved DeployTOIss tests to check the actual website and web api deployment state; added upgrade db test
