+++ source/Esf.DataAccess.DeploymentTests/Esf.DataAccess.DeploymentTests.csproj
@@ -100,7 +100,8 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="EsStatesRepositoryTests.cs" />
+    <Compile Include="EsFiddleDeploymentTests.cs" />
+    <Compile Include="UpgradeScriptAuditRecord.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
