+++ source/Esf.Deployment/DeployToIss.tests.ps1
@@ -20,36 +20,61 @@ Describe "esf.WebApi deployment"  {
 	Context "deployment on IIS" {		
 		$esfWebApiConfig = GetDeploymentConfig("$PSScriptRoot\esfWebApi.config.json")
 		ImportWebAdministrationModule
+		$esfWebApi = Get-Website -Name "esf.WebApi"
+
 		It "should have output path in config.json" {
-			$esfWebApiConfig.DeploymentPath | Should Be "..\Esf.WebApi"
+			$deploymentPath = $esfWebApiConfig.DeploymentPath
+			$expectedPath = Resolve-Path $PSScriptRoot\$deploymentPath
+			$esfWebApi.PhysicalPath | Should Be $expectedPath.Path
 		}
 
 		It "should have port number in config.json" {
-			$esfWebApiConfig.Port | Should Be 40081
+			$expectedPort = $esfWebApiConfig.Port
+			$actualBinding = $esfWebApi.Bindings.Collection.BindingInformation
+			$actualBinding | Should Match $expectedPort
 		}
 
 		It "should have deployed and started website" {
-			$esfWebApi = Get-Website -Name "esf.WebApi"
+			
 			$esfWebApi.State | Should Be "Started"
+		}
+
+		It "should return states/new resource" {
+			$portNumber = $esfWebApiConfig.Port
+			$NavigationHeadResponse = Invoke-WebRequest -Uri "http://localhost:$portNumber/states/new/" -Method Get
+			$NavigationHeadResponse.StatusCode | Should Be 200
+			$NavigationHeadResponse.StatusDescription | Should Be "OK"
 		}
 	}
 }
 
 Describe "esf.Website IIS deployment" {
 	Context "deployment on IIS" {		
-		$esfWebApiConfig = GetDeploymentConfig("$PSScriptRoot\esfWebsite.config.json")
+		$esfWebsiteConfig = GetDeploymentConfig("$PSScriptRoot\esfWebsite.config.json")
 		ImportWebAdministrationModule
-		It "should have output path in config.json" {
-			$esfWebApiConfig.DeploymentPath | Should Be "..\Esf.Website"
+		$esfWebsite = Get-Website -Name "esf.Website"
+
+		It "should have the right deployment path" {
+			$deploymentPath = $esfWebsiteConfig.DeploymentPath
+			$expectedPath = Resolve-Path $PSScriptRoot\$deploymentPath
+			$esfWebsite.PhysicalPath | Should Be $expectedPath.Path
 		}
 
-		It "should have port number in config.json" {
-			$esfWebApiConfig.Port | Should Be 40082
+		It "should have the right port number" {
+			$expectedPort = $esfWebsiteConfig.Port
+			$actualBinding = $esfWebsite.Bindings.Collection.BindingInformation
+			$actualBinding | Should Match $expectedPort
 		}
 
-		It "should have deployed and started website" {
-			$esfWebApi = Get-Website -Name "esf.Website"
-			$esfWebApi.State | Should Be "Started"
+		It "should be started" {			
+			$esfWebsite.State | Should Be "Started"
+		}
+
+		It "should be able to navigate the home page" {
+			$portNumber = $esfWebsiteConfig.Port
+			$NavigationHeadResponse = Invoke-WebRequest -Uri "http://localhost:$portNumber" -Method Head
+			$NavigationHeadResponse.StatusCode | Should Be 200
+			$NavigationHeadResponse.StatusDescription | Should Be "OK"
 		}
 	}
 }
